Move SetProperty to package cfg
